show correct number of matches
